pyproject.toml
[project]
name = "burmese-money-scam-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "emoji>=2.15.0",
    "gensim>=4.3.3",
    "mmdt-tokenizer",
    "nltk>=3.9.1",
    "openpyxl>=3.1.5",
    "pandas>=2.0.3",
    "scikit-learn>=1.3.2",
    "skops>=0.9.0",
    "tabulate>=0.9.0",
    "torch>=2.5.1",
    "xgboost>=2.1.4",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.30.0",
    "mlflow>=2.14.0",
    "matplotlib>=3.10.0",
    "pyyaml>=6.0.1",
]

[tool.uv.sources]
mmdt-tokenizer = { workspace = true }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "plotly>=6.5.0",
]
